SOAP is now functional
Create instances of SOAP_Client with a URL and array of namespaces.

Invoke a RPC using call()

XMLObject::with_namespace() has been updated to support an optional
namespace, if not specified defaults to $this->namespace. This is to
support the new own_namespace property, which if set will use a
different name space for the base object than its properties.

In serialize_writer we don't need to call startElement and endElement
when we are dealing with other XMLObject's as they will call it during
the call to XMLObject::create_xml()

More documentation.